change commands.py to json.py module
