+++ python/pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 markers =
     smoke_test: mark a test as a build verification testing.
-addopts = -k "not test_redis_modules.py and not test_json.py"
+addopts = --ignore "tests_redis_modules"
