--- python/python/glide/async_commands/redis_modules/json/__init__.py
@@ -1,5 +1,0 @@
-# Copyright GLIDE-for-Redis Project Contributors - SPDX Identifier: Apache-2.0
-
-from .commands import ConditionalChange, Json
-
-__all__ = ["ConditionalChange", "Json"]
