+++ python/python/tests/tests_redis_modules/test_json.py
@@ -4,7 +4,7 @@ import json as OuterJson
 
 import pytest
 from glide.async_commands.core import ConditionalChange, InfoSection
-from glide.async_commands.redis_modules.json.commands import Json
+from glide.async_commands.redis_modules.json import json
 from glide.config import ProtocolVersion
 from glide.constants import OK
 from glide.redis_client import TRedisClient
@@ -22,10 +22,12 @@ class TestJson:
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_json_set_get(self, redis_client: TRedisClient):
-        json = Json()
         key = get_random_string(5)
 
-        assert await json.set(redis_client, key, "$", {"a": 1.0, "b": 2}) == OK
+        assert (
+            await json.set(redis_client, key, "$", OuterJson.dumps({"a": 1.0, "b": 2}))
+            == OK
+        )
 
         result = await json.get(redis_client, key, "$")
         assert isinstance(result, str)
@@ -35,18 +37,20 @@ class TestJson:
         assert isinstance(result, str)
         assert OuterJson.loads(result) == {"$.a": [1.0], "$.b": [2]}
 
-        assert await json.get(redis_client, "non_existing_key", "$", "bar") == None
+        assert await json.get(redis_client, "non_existing_key", "$", "bar") is None
         assert await json.get(redis_client, key, "$.d") == "[]"
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_json_set_get_multiple_values(self, redis_client: TRedisClient):
-        json = Json()
         key = get_random_string(5)
 
         assert (
             await json.set(
-                redis_client, key, "$", {"a": {"c": 1, "d": 4}, "b": {"c": 2}, "c": 3}
+                redis_client,
+                key,
+                "$",
+                OuterJson.dumps({"a": {"c": 1, "d": 4}, "b": {"c": 2}, "c": 3}),
             )
             == OK
         )
@@ -59,7 +63,7 @@ class TestJson:
         assert isinstance(result, str)
         assert OuterJson.loads(result) == {"$..c": [3, 1, 2], "$.c": [3]}
 
-        assert await json.set(redis_client, key, "$..c", "new_value") == OK
+        assert await json.set(redis_client, key, "$..c", '"new_value"') == OK
         result = await json.get(redis_client, key, "$..c")
         assert isinstance(result, str)
         assert OuterJson.loads(result) == ["new_value"] * 3
@@ -67,24 +71,24 @@ class TestJson:
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_json_set_conditional_set(self, redis_client: TRedisClient):
-        json = Json()
         key = get_random_string(5)
+        value = OuterJson.dumps({"a": 1.0, "b": 2})
         assert (
             await json.set(
                 redis_client,
                 key,
                 "$",
-                {"a": 1.0, "b": 2},
+                value,
                 ConditionalChange.ONLY_IF_EXISTS,
             )
-            == None
+            is None
         )
         assert (
             await json.set(
                 redis_client,
                 key,
                 "$",
-                {"a": 1.0, "b": 2},
+                value,
                 ConditionalChange.ONLY_IF_DOES_NOT_EXIST,
             )
             == OK
@@ -95,10 +99,10 @@ class TestJson:
                 redis_client,
                 key,
                 "$.a",
-                4.5,
+                "4.5",
                 ConditionalChange.ONLY_IF_DOES_NOT_EXIST,
             )
-            == None
+            is None
         )
 
         assert await json.get(redis_client, key, ".a") == "1.0"
@@ -108,7 +112,7 @@ class TestJson:
                 redis_client,
                 key,
                 "$.a",
-                4.5,
+                "4.5",
                 ConditionalChange.ONLY_IF_EXISTS,
             )
             == OK
@@ -119,11 +123,13 @@ class TestJson:
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_json_get_formatting(self, redis_client: TRedisClient):
-        json = Json()
         key = get_random_string(5)
         assert (
             await json.set(
-                redis_client, key, "$", {"a": 1.0, "b": 2, "c": {"d": 3, "e": 4}}
+                redis_client,
+                key,
+                "$",
+                OuterJson.dumps({"a": 1.0, "b": 2, "c": {"d": 3, "e": 4}}),
             )
             == OK
         )
